- module 구조 변경  : d ->e, e -> f  : d 분석모듈로 추가
