+++ src/check_d.py
@@ -1,51 +1,45 @@
 import json
 import sys
-
-def warn(msg):
-    print(f"[WARN] {msg}")
+import os
 
 def fail(msg):
     print(f"[ERROR] {msg}")
     sys.exit(1)
 
 def main():
-    path = "outputs/biz_opportunities.json"
+    # 1. 분석 요약 파일 확인
+    if not os.path.exists("outputs/analysis_summary.json"):
+        fail("outputs/analysis_summary.json 없음")
+    
     try:
-        with open(path, "r", encoding="utf-8") as f:
-            data = json.load(f)
+        with open("outputs/analysis_summary.json", "r", encoding="utf-8") as f:
+            summary = json.load(f)
     except Exception:
-        fail("biz_opportunities.json 로드 실패")
-
-    ideas = data.get("ideas")
-    if ideas is None or not isinstance(ideas, list):
-        fail("ideas 필드가 누락되었거나 형식이 잘못됨")
-
-    n = len(ideas)
-    if n == 0:
-        warn("아이디어가 0개입니다(폴백 비활성화 설정).")
-    elif n < 3:
-        warn(f"아이디어가 적습니다({n}개).")
-
-    # 필드(축소 버전)
-    required = [
-        "idea",
-        "problem",
-        "target_customer",
-        "value_prop",
-        "solution",
-        "risks",
-        "priority_score",
-    ]
-
-    # 앞쪽 최대 5건 샘플 검증(있을 때만)
-    for idx, it in enumerate(ideas[:5], 1):
-        if not isinstance(it, dict):
-            fail(f"아이디어 형식 오류(index={idx})")
-        for k in required:
-            if k not in it:
-                fail(f"필드 누락: {k} (index={idx})")
-
-    print(f"[INFO] Check D OK | ideas={n}")
+        fail("analysis_summary.json 로드 실패")
+    
+    # 2. 매트릭스 파일 확인
+    if not os.path.exists("outputs/export/company_topic_matrix_wide.csv"):
+        fail("company_topic_matrix_wide.csv 없음")
+    
+    if not os.path.exists("outputs/export/company_topic_matrix_long.csv"):
+        fail("company_topic_matrix_long.csv 없음")
+    
+    # 3. 네트워크 파일 확인
+    if not os.path.exists("outputs/company_network.json"):
+        fail("company_network.json 없음")
+    
+    try:
+        with open("outputs/company_network.json", "r", encoding="utf-8") as f:
+            network = json.load(f)
+    except Exception:
+        fail("company_network.json 로드 실패")
+    
+    # 4. 기본 검증
+    if "edges" not in network:
+        fail("company_network.json에 edges 필드 없음")
+    
+    print(f"[INFO] Check D OK | matrix_orgs={summary.get('matrix_stats', {}).get('num_orgs', 0)} "
+          f"network_edges={summary.get('network_stats', {}).get('num_edges', 0)}")
 
 if __name__ == "__main__":
     main()
